Accommodates stop failing when mysqld is already down.
Fixes #3 by adding a test to the stop command to avoid running stop when mysqld is already stopped. This accommodates the mode of operations of the init script's stop command on older versions of MySQL on Redhat/CentOS.

Also added additional tests to make sure the commands work from a number of different starting states.